feat(deploy): 由 Pages 改为 Worker，接通内网 PostgreSQL
- wrangler.toml：main = functions-worker/index.js + [assets]（SPA 回退 + run_worker_first），
  域名改用 zone route（custom_domain 会因已有 Pages DNS 记录报 100117），仅保留 ieop.top
- package.json：build = vite build + wrangler pages functions build；新增 deploy
- .gitignore：忽略 functions-worker/ 构建产物
- functions/lib/db.js + scripts/pg-schema.sql：builtin_sites.rate 改 BIGINT
  （线上实测 901001003538，超出 int4，导入会报 integer out of range）
- DEPLOY.md：部署流程改写为 Worker 形态，新增「十一、从 Pages 迁移到 Worker」记录四个坑

+++ functions/lib/db.js
@@ -210,7 +210,7 @@ export async function upsertBuiltinSites(env, rows) {
   if (!rows?.length) return
   const SQL = `INSERT INTO ${T('builtin_sites')} (url, name, icon, icon_src, description, rate, source_id, updated_at)
      SELECT x.url, x.name, x.icon, x."iconSrc", x."description",
-            x.rate::integer, x."sourceId", x."updatedAt"
+            x.rate::bigint, x."sourceId", x."updatedAt"
        FROM jsonb_to_recordset($1::jsonb) AS x(url text, name text, icon text, "iconSrc" text, "description" text, rate text, "sourceId" text, "updatedAt" text)
      WHERE true
      ON CONFLICT (url) DO UPDATE SET
